Adds HGO_4Fibers anisotropic model
Adds the HGO_4Fibers anisotropic hyperelastic constitutive model, and enables composition of mechanical models using the '+'
operator.

This commit introduces a new anisotropic hyperelastic material model, HGO_4Fibers, which accounts for the presence of four families of fibers within the material. Additionally, it overloads the '+' operator to allow for easy composition of mechanical models, such as combining isotropic and anisotropic components.  This simplifies the creation of complex material behaviors.

New AnisoElastic and IsoElastic abstract types introduced